add publishing to broccoli branch
This changes the url that backburner was previously posted to from:

http://builds.emberjs.com.s3.amazonaws.com/backburner.js/lastest/backburner.js

to

http://builds.emberjs.com.s3.amazonaws.com/backburner.js/backburner-latest.js
AND
http://builds.emberjs.com.s3.amazonaws.com/backburner.js/lastest/backburner-min.js
AND
http://builds.emberjs.com.s3.amazonaws.com/backburner.js/lastest/backburner-amd.js

This is consistent with RSVP's style which is the most closely related
micro lib (as far as build is concerned) to backburner. If this is
undesired we can add the other urls

IMPORTANT

In order for the publishing to work the following variables have to be
set

S3_BUCKET_NAME
S3_SECRET_ACCESS_KEY
S3_ACCESS_KEY_ID

inside the .travis.yml. I am unable to set this at the moment, so this
is something that should be kept in mind during merge.